requirements.txt
addict==2.4.0
aiofiles==24.1.0
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
attrs==25.4.0
av==16.0.1
axial_positional_embedding==0.3.12
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy @ file:///croot/contourpy_1732540045555/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_debugpy_1758162044/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
decord==0.6.0
dotenv==0.9.9
easydict==1.13
einops==0.8.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fastapi==0.121.1
ffmpy==1.0.0
filelock==3.20.0
fire==0.7.1
flatbuffers==25.9.23
fonttools @ file:///home/task_175975444451365/conda-bld/fonttools_1759754870571/work
fsspec==2025.10.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.45
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hyper-connections==0.2.1
idna==3.11
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
iopath==0.1.10
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipykernel_1761567932/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1748711175/work
iterative-stratification==0.1.9
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_core_1760643864/work
kiwisolver @ file:///croot/kiwisolver_1737039087198/work
llvmlite==0.46.0
local-attention @ git+https://github.com/lucidrains/local-attention.git@af58e53f81e5e1d87a7d50cb193d5c26f3c0e03f
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mdurl==0.1.2
mkl-service==2.5.2
mkl_fft @ file:///home/task_176159252575002/conda-bld/mkl_fft_1761592901585/work
mkl_random @ file:///home/task_176159293813909/conda-bld/mkl_random_1761592952512/work
ml_dtypes==0.5.4
mmengine==0.10.7
mpmath==1.3.0
narwhals==2.11.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
numba==0.63.1
numpy @ file:///home/task_176157872663408/conda-bld/numpy_and_numpy_base_1761579342634/work/dist/numpy-2.2.5-cp310-cp310-linux_x86_64.whl#sha256=fb5b85b589f86e6404be6a37347b98be70b32c4ae837bb26e1180e9f9a7fb718
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnx==1.19.1
onnx-ir==0.1.12
onnxruntime==1.23.2
onnxruntime-gpu==1.23.2
onnxscript==0.5.6
onnxslim==0.1.75
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
orjson==3.11.4
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.3
parameterized==0.9.0
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
performer-pytorch==1.1.4
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1759953252/work
polars==1.35.2
polars-runtime-32==1.35.2
portalocker==3.2.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
protobuf==6.33.0
psutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_psutil_1762092922/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==21.0.0
pybboxes==0.1.6
pycocotools==2.0.10
pydantic==2.11.10
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pynndescent==0.5.13
pyparsing @ file:///croot/pyparsing_1731445506121/work
PyQt6==6.9.1
PyQt6_sip @ file:///home/task_175751008770153/conda-bld/pyqt-split_1757511022671/work/pyqt_sip
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.2.1
python-multipart==0.0.20
pytorchvideo==0.1.5
pytz==2025.2
PyYAML==6.0.3
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyzmq_1757387023/work
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
ruff==0.14.4
safehttpx==0.1.7
safetensors==0.6.2
sahi==0.11.36
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.44.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.3
sip @ file:///croot/sip_1756223371714/work
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap==5.0.2
sniffio==1.3.1
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.49.3
streamlit==1.51.0
stringzilla==4.2.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.2.0
terminaltables==3.1.10
threadpoolctl==3.6.0
timm==0.5.4
tokenizers==0.22.1
toml==0.10.2
tomli @ file:///croot/tomli_1753774587605/work
tomlkit==0.13.3
torch==2.9.1+cu128
torchinfo==1.8.0
torchvision==0.24.1+cu128
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1762506714284/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.57.1
triton==3.5.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
ultralytics==8.3.231
ultralytics-thop==2.0.18
umap==0.1.1
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.38.0
wandb==0.23.0
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
websockets==15.0.1
yacs==0.1.8
yapf==0.43.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
